backend/requirements.txt
# ShuttleSense Backend Dependencies
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Computer Vision & ML
mediapipe==0.10.9
opencv-python-headless==4.9.0.80
numpy==1.26.3
scipy==1.12.0

# AI/LLM
google-generativeai==0.3.2

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# Configuration
python-dotenv==1.0.0

# Async Support
aiofiles==23.2.1

# Logging (structured)
python-json-logger==2.0.7

# System Monitoring
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Database (for future migration)
# sqlalchemy==2.0.25
# asyncpg==0.29.0
# alembic==1.13.1

# Task Queue (for async processing - future)
# celery==5.3.6
# redis==5.0.1
